GET random_merchant_pos_new_checkout_request
