+++ codebase/services/trigger/service.py
@@ -15,13 +15,5 @@ def api():
             None, 
             handle_trigger_random_merchant_pos_new_checkout_request
         )()
-    
-    # @api.get("/")
-    # def get_merchant_pos_new_checkout():
-    #     return request_handler(
-    #         definition,
-    #         NullRequest, 
-    #         handle_trigger_merchant_pos_new_checkout_request
-    #     )(None)
-    
+        
     return api
