+++ trigger-random-new-merchant-pos-checkout
